feat: `project.items.add(contentId, fields)`
