+++ api/items.add.js
@@ -3,6 +3,7 @@
 import { addIssueToProjectMutation } from "./lib/queries.js";
 import { projectItemNodeToGitHubProjectItem } from "./lib/project-item-node-to-github-project-item.js";
 import { getStateWithProjectFields } from "./lib/get-state-with-project-fields.js";
+import { getFieldsUpdateQuery } from "./lib/get-fields-update-query.js";
 
 /**
  * Adds new item to project. Loads project fields and caches them unless already loaded.
@@ -10,9 +11,10 @@ import { getStateWithProjectFields } from "./lib/get-state-with-project-fields.j
  * @param {import("../").default} project
  * @param {import("..").GitHubProjectState} state
  * @param {string} contentNodeId
+ * @param {Record<string, string>} fields
  * @returns {Promise<import("..").GitHubProjectItem>}
  */
-export default async function addItem(project, state, contentNodeId) {
+export default async function addItem(project, state, contentNodeId, fields) {
   const stateWithFields = await getStateWithProjectFields(project, state);
 
   const {
@@ -22,5 +24,24 @@ export default async function addItem(project, state, contentNodeId) {
     contentId: contentNodeId,
   });
 
-  return projectItemNodeToGitHubProjectItem(stateWithFields, projectNextItem);
+  const newItem = projectItemNodeToGitHubProjectItem(
+    stateWithFields,
+    projectNextItem
+  );
+
+  if (!fields) return newItem;
+
+  const query = getFieldsUpdateQuery(stateWithFields, fields);
+  await project.octokit.graphql(query, {
+    projectId: stateWithFields.id,
+    itemId: newItem.id,
+  });
+
+  return {
+    ...newItem,
+    fields: {
+      ...newItem.fields,
+      ...fields,
+    },
+  };
 }
